pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-tools"
version = "0.1.0"
description = "AI Tools for text processing using LLMs"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.1.0",
    "httpx>=0.28.1",
    "langchain-openai>=1.0.1",
    "langgraph>=0.2.0",
    "oci>=2.161.1",
    "oci-openai>=0.1.0",
    "openai>=2.6.0",
    "pydantic>=2.12.3",
    "pydantic-settings[yaml]>=2.0.0",
    "pyyaml>=6.0.0",
    "envyaml>=1.10.211231",
    "pytest>=7.0.0",
]

[project.scripts]
ai_tool = "clients.multi_tool_client:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_tools", "examples", "clients"]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
]
